avm1: Merge Activation into StackFrame
